Рефакторинг: модули перемещены в отдельные директории; функция getElement переименована в getElementFromTemplate; улучшен код в функции showScreen.
