--- js/get-element.js
@@ -1,7 +1,0 @@
-const getElement = (template) => {
-  const element = document.createElement(`template`);
-  element.innerHTML = template;
-  return element;
-};
-
-export default getElement;
