+++ js/main.js
@@ -1,12 +1,12 @@
-import intro from './intro';
-import showScreen from './show-screen';
-import greeting from './greeting';
+import showScreen from './utils/show-screen';
+import intro from './screens/intro';
+import greeting from './screens/greeting';
 
 document.onload = function () {
   showScreen(intro);
 };
 
-document.querySelector(`main`).addEventListener(`click`, function (event) {
+document.querySelector(`body`).addEventListener(`click`, function (event) {
   const target = event.target;
   if (!target.parentNode.classList.contains(`back`)) {
     return;
