+++ js/screens/game-2.js
@@ -1,6 +1,7 @@
-import getElement from './get-element';
-import stats from './stats';
-import showScreen from './show-screen';
+import getElementFromTemplate from '../utils/get-element-from-template';
+import showScreen from '../utils/show-screen';
+import answerIsSelected from '../utils/check-answers';
+import game3 from './game-3';
 
 const template = `<header class="header">
 <div class="header__back">
@@ -17,16 +18,18 @@ const template = `<header class="header">
 </div>
 </header>
 <div class="game">
-<p class="game__task">Найдите рисунок среди изображений</p>
-<form class="game__content  game__content--triple">
+<p class="game__task">Угадай, фото или рисунок?</p>
+<form class="game__content  game__content--wide">
   <div class="game__option">
-    <img src="http://placehold.it/304x455" alt="Option 1" width="304" height="455">
-  </div>
-  <div class="game__option  game__option--selected">
-    <img src="http://placehold.it/304x455" alt="Option 1" width="304" height="455">
-  </div>
-  <div class="game__option">
-    <img src="http://placehold.it/304x455" alt="Option 1" width="304" height="455">
+    <img src="http://placehold.it/705x455" alt="Option 1" width="705" height="455">
+    <label class="game__answer  game__answer--photo">
+      <input name="question1" type="radio" value="photo">
+      <span>Фото</span>
+    </label>
+    <label class="game__answer  game__answer--wide  game__answer--paint">
+      <input name="question1" type="radio" value="paint">
+      <span>Рисунок</span>
+    </label>
   </div>
 </form>
 <div class="stats">
@@ -55,21 +58,21 @@ const template = `<header class="header">
 </div>
 </footer>`;
 
-const game3 = getElement(template);
+const game2 = getElementFromTemplate(template);
 
-document.querySelector(`main`).addEventListener(`click`, function (event) {
-  const main = document.querySelector(`main`);
-  if (main.dataset.game !== `3`) {
+document.querySelector(`body`).addEventListener(`click`, function (event) {
+  const body = document.querySelector(`body`);
+  if (body.dataset.game !== `2`) {
     return;
   }
-  let target = event.target;
-  while (target !== null && target !== main) {
-    if (target.classList.contains(`game__option`)) {
-      showScreen(stats);
-      return;
-    }
-    target = target.parentNode;
+  const target = event.target;
+  if (target.type !== `radio`) {
+    return;
+  }
+  if (answerIsSelected(document.querySelectorAll(`input[name=question1]`))) {
+    showScreen(game3);
+    body.dataset.game = `3`;
   }
 });
 
-export default game3;
+export default game2;
