--- js/show-screen.js
@@ -1,5 +1,0 @@
-const showScreen = (element) => {
-  document.querySelector(`main`).innerHTML = element.innerHTML;
-};
-
-export default showScreen;
