+++ js/utils/get-element-from-template.js
@@ -1,0 +1,8 @@
+const getElementFromTemplate = (template) => {
+  const element = document.createElement(`main`);
+  element.classList.add(`central`);
+  element.innerHTML = template;
+  return element;
+};
+
+export default getElementFromTemplate;
