+++ js/utils/show-screen.js
@@ -1,0 +1,6 @@
+const showScreen = (element) => {
+  const mainElement = document.querySelector(`main`);
+  mainElement.parentNode.replaceChild(element, mainElement);
+};
+
+export default showScreen;
